+++ src/components/home/Result/Result.tsx
