+++ src/components/home/Result/ResultItem.tsx
