+++ src/components/home/Result/index.ts
@@ -1,0 +1,1 @@
+export { default as Result } from './Result';
