+++ src/components/home/index.ts
@@ -1,0 +1,3 @@
+export * from './Result';
+export { default as Form } from './Form';
+export { default as Header } from './Header';
