+++ src/components/index.ts
@@ -1,2 +1,2 @@
-export * from './InputField';
+export * from './home/InputField';
 export * from './layouts';
